requirements.txt
flask
Flask-SQLAlchemy
requests
pytest
pytest-cov
